[Refactor] Modularize multi-SCM plain-issue source-control helpers (#664)
